backend/requirements.txt
asgiref==3.9.1
babel==2.17.0
boto3==1.40.6
botocore==1.40.6
Django==5.2.5
django-debug-toolbar==6.0.0
django-extensions==4.1
django-money==3.5.4
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
psycopg2-binary==2.9.10
py-moneyed==2.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
rest-framework-money-field==2.2.1
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
